fix: resolve clippy warnings and improve code quality
Resolved multiple clippy warnings including:
- Fixed format string issues to use inline format arguments
- Fixed unnecessary casting in DOS protection monitoring
- Removed needless reference operations in encryption
- Fixed clone-on-copy issues in message handling
- Replaced redundant pattern matching with is_err()
- Fixed await-holding-lock issue in transport layer
- Replaced or_insert_with with or_default where appropriate
- Updated NonZeroU32 constants to use safe construction
- Fixed trim-before-split-whitespace pattern
- Replaced get(0) with first() accessor

Code is now cleaner and follows Rust best practices more closely
while maintaining full functionality and test coverage.

+++ multivm-p2p/src/security/dos_protection.rs
@@ -626,7 +626,7 @@ impl DosProtectionManager {
 
         // Memory usage - get resident memory in MB
         monitor.memory_usage_mb = if let Some(proc) = process {
-            ((proc.memory() / 1024 / 1024) as u64) as usize
+            (proc.memory() / 1024 / 1024) as usize
         } else {
             // Fallback to system memory usage
             (((sys.used_memory() / 1024 / 1024) as f64 * 0.1) as u64) as usize // Estimate 10% of system memory
